Extract common table data types to Zebra.Table.Data
